Each unit will have individual idle timer
