+++ src/fheroes2/battle/battle_troop.h
@@ -31,6 +31,7 @@
 #include "battle_arena.h"
 #include "battle_army.h"
 #include "bitmodes.h"
+#include "game_delays.h"
 #include "sprite.h"
 
 class Spell;
@@ -186,6 +187,7 @@ namespace Battle
 
         int GetAnimationState() const;
         bool isIdling() const;
+        bool checkIdleDelay();
 
         // Find a better way to expose it without a million getters/setters
         AnimationState animation;
@@ -205,6 +207,8 @@ namespace Battle
         Position position;
         ModesAffected affected;
         Unit * mirror;
+        TimeDelay idleTimer;
+        bool idleTimerSet;
 
         // These variables are mutable due to population of them of the fly as we don't want to calculate everything
         mutable std::map<int, Surface> contoursMain;
