+++ src/fheroes2/game/game_delays.h
@@ -1,0 +1,61 @@
+/***************************************************************************
+ *   Free Heroes of Might and Magic II: https://github.com/ihhub/fheroes2  *
+ *   Copyright (C) 2020                                                    *
+ *                                                                         *
+ *   This program is free software; you can redistribute it and/or modify  *
+ *   it under the terms of the GNU General Public License as published by  *
+ *   the Free Software Foundation; either version 2 of the License, or     *
+ *   (at your option) any later version.                                   *
+ *                                                                         *
+ *   This program is distributed in the hope that it will be useful,       *
+ *   but WITHOUT ANY WARRANTY; without even the implied warranty of        *
+ *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the         *
+ *   GNU General Public License for more details.                          *
+ *                                                                         *
+ *   You should have received a copy of the GNU General Public License     *
+ *   along with this program; if not, write to the                         *
+ *   Free Software Foundation, Inc.,                                       *
+ *   59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.             *
+ ***************************************************************************/
+
+#ifndef H2GAME_DELAYS_H
+#define H2GAME_DELAYS_H
+
+#include <thread.h>
+
+struct TimeDelay : std::pair<SDL::Time, uint32_t>
+{
+    TimeDelay( uint32_t dl )
+    {
+        second = dl;
+    }
+
+    uint32_t operator()( void ) const
+    {
+        return second;
+    }
+
+    TimeDelay & operator=( uint32_t dl )
+    {
+        second = dl;
+        return *this;
+    }
+
+    void Reset( void )
+    {
+        first.Start();
+    }
+
+    bool Trigger( uint32_t customDelay = 0 )
+    {
+        first.Stop();
+        const uint32_t expected = ( customDelay > 0 ) ? customDelay : second;
+        if ( first.Get() < expected )
+            return false;
+
+        first.Start();
+        return true;
+    }
+};
+
+#endif
